requirements.txt
gunicorn
flask
firebase-admin
replicate
pypng
urllib3
pyqrcode
